✨ Extract reusable SidebarShell component
Refactors the 672-line monolithic Sidebar into a configurable SidebarShell
that can be reused across contexts (main, enterprise, white-label).

SidebarShell accepts:
- navSections: configurable navigation items
- features: toggle missions, studio, cluster status, etc.
- branding: title/logo for white-label
- All visual features: collapse, pin, resize, mobile, glass effect

Both Sidebar and EnterpriseSidebar now compose SidebarShell.
Enterprise sidebar gains AI missions and Console Studio support.
