+++ web/src/components/layout/Sidebar.tsx
@@ -1,672 +1,68 @@
-import { useState, useRef, useEffect, useCallback } from 'react'
-import { useTranslation } from 'react-i18next'
-import { NavLink, useNavigate, useLocation } from 'react-router-dom'
-import { Plus, ChevronLeft, ChevronRight, CheckCircle2, AlertTriangle, WifiOff, GripVertical, X, User, Pin, PinOff, Satellite, Loader2 } from 'lucide-react'
-import { iconRegistry } from '../../lib/icons'
-import { cn } from '../../lib/cn'
-import { Tooltip } from '../ui/Tooltip'
-import { SnoozedCards } from './SnoozedCards'
-import { useSidebarConfig, SidebarItem, PROTECTED_SIDEBAR_IDS, SIDEBAR_COLLAPSED_WIDTH_PX, SIDEBAR_DEFAULT_WIDTH_PX } from '../../hooks/useSidebarConfig'
-import { useMobile } from '../../hooks/useMobile'
-import { useClusters } from '../../hooks/mcp/clusters'
-import { isClusterUnreachable, isClusterHealthy } from '../clusters/utils'
+import { useLocation, useNavigate } from 'react-router-dom'
+import { SidebarShell } from './SidebarShell'
+import type { NavSection, SidebarNavItem } from './SidebarShell'
+import { useSidebarConfig } from '../../hooks/useSidebarConfig'
+import type { SidebarItem } from '../../hooks/useSidebarConfig'
 import { useDashboardContextOptional } from '../../hooks/useDashboardContext'
-import type { SnoozedSwap } from '../../hooks/useSnoozedCards'
-import type { SnoozedRecommendation } from '../../hooks/useSnoozedRecommendations'
-import type { SnoozedMission } from '../../hooks/useSnoozedMissions'
-import { useActiveUsers } from '../../hooks/useActiveUsers'
-import { useMissions } from '../../hooks/useMissions'
 import { ROUTES } from '../../config/routes'
-import { DASHBOARD_CONFIGS } from '../../config/dashboards/index'
-import { emitSidebarNavigated, emitDashboardRenamed } from '../../lib/analytics'
-import { prefetchDashboard } from '../../lib/prefetchDashboard'
-import { useVersionCheck } from '../../hooks/useVersionCheck'
-import { useUpgradeState } from '../../hooks/useUpgradeState'
-import { STORAGE_KEY_GROUND_CONTROL_DASHBOARDS } from '../../lib/constants/storage'
-import { SIDEBAR_CONTROLS_LEFT_OFFSET_PX } from '../../lib/constants/ui'
-import { safeGetJSON } from '../../lib/utils/localStorage'
 
-/** Sidebar resize limits in pixels */
-const SIDEBAR_MIN_WIDTH_PX = 180
-const SIDEBAR_MAX_WIDTH_PX = 480
-
-/** Map sidebar item href to dashboard config ID for card count display */
-const HREF_TO_DASHBOARD_ID: Record<string, string> = {
-  '/': 'main', '/compute': 'compute', '/security': 'security',
-  '/gitops': 'gitops', '/storage': 'storage', '/network': 'network',
-  '/events': 'events', '/workloads': 'workloads', '/operators': 'operators',
-  '/clusters': 'clusters', '/compliance': 'compliance', '/cost': 'cost',
-  '/gpu-reservations': 'gpu', '/nodes': 'nodes', '/deployments': 'deployments',
-  '/pods': 'pods', '/services': 'services', '/helm': 'helm',
-  '/alerts': 'alerts', '/ai-ml': 'ai-ml', '/ci-cd': 'ci-cd',
-  '/logs': 'logs', '/data-compliance': 'data-compliance', '/arcade': 'arcade',
-  '/deploy': 'deploy', '/ai-agents': 'ai-agents',
-  '/llm-d-benchmarks': 'llm-d-benchmarks', '/cluster-admin': 'cluster-admin',
-  '/insights': 'insights', '/drasi': 'drasi',
-  '/multi-tenancy': 'multi-tenancy', '/acmm': 'acmm' }
-
-/** Prefix for custom dashboard routes */
-const CUSTOM_DASHBOARD_PREFIX = '/custom-dashboard/'
-
-/** Check if a sidebar item's href points to a Ground Control dashboard */
-function isGroundControlItem(href: string): boolean {
-  if (!href.startsWith(CUSTOM_DASHBOARD_PREFIX)) return false
-  const dashboardId = href.slice(CUSTOM_DASHBOARD_PREFIX.length)
-  const gcMapping = safeGetJSON<Record<string, unknown>>(STORAGE_KEY_GROUND_CONTROL_DASHBOARDS) ?? {}
-  return dashboardId in gcMapping
+/** Convert a SidebarItem from useSidebarConfig into SidebarShell's NavItem format */
+function toNavItem(item: SidebarItem): SidebarNavItem {
+  return {
+    id: item.id,
+    label: item.name,
+    href: item.href,
+    icon: item.icon,
+    isCustom: item.isCustom,
+  }
 }
 
 export function Sidebar() {
-  const { config, toggleCollapsed, setCollapsed, reorderItems, updateItem, removeItem, closeMobileSidebar, setWidth } = useSidebarConfig()
-  const { isMobile } = useMobile()
-  const { deduplicatedClusters } = useClusters()
+  const { config } = useSidebarConfig()
   const dashboardContext = useDashboardContextOptional()
-  const { isFullScreen: isMissionFullScreen } = useMissions()
-  const { viewerCount, hasError: viewersError, isLoading: viewersLoading } = useActiveUsers()
-  const { hasUpdate, channel, latestMainSHA } = useVersionCheck()
-  const upgradeState = useUpgradeState()
-  const isUpgrading = upgradeState.phase === 'triggering' || upgradeState.phase === 'restarting'
-  const { t } = useTranslation()
   const navigate = useNavigate()
   const location = useLocation()
 
-  // Close mobile sidebar on route change
-  useEffect(() => {
-    if (isMobile) {
-      closeMobileSidebar()
-    }
-  }, [location.pathname, isMobile, closeMobileSidebar])
-
-  // Auto-hide: collapse sidebar when mouse leaves, expand on hover
-  const SIDEBAR_AUTO_HIDE_MS = 2000
-  const [isPinned, setIsPinned] = useState(() => {
-    try { return localStorage.getItem('sidebar-left-pinned') !== 'false' } catch { return true }
-  })
-  const autoHideTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-  const clearAutoHideTimer = useCallback(() => {
-    if (autoHideTimerRef.current) {
-      clearTimeout(autoHideTimerRef.current)
-      autoHideTimerRef.current = null
-    }
-  }, [])
-
-  const handleSidebarMouseEnter = () => {
-    clearAutoHideTimer()
-    if (!isPinned && config.collapsed && !isMobile) {
-      setCollapsed(false)
-    }
-  }
-
-  const handleSidebarMouseLeave = () => {
-    if (!isPinned && !isMobile) {
-      clearAutoHideTimer()
-      autoHideTimerRef.current = setTimeout(() => {
-        setCollapsed(true)
-      }, SIDEBAR_AUTO_HIDE_MS)
-    }
-  }
-
-  const toggleSidebarPin = () => {
-    setIsPinned(prev => {
-      const next = !prev
-      try { localStorage.setItem('sidebar-left-pinned', String(next)) } catch { /* ignore */ }
-      if (next) {
-        clearAutoHideTimer()
-        // Pinning while collapsed should expand the sidebar (#8647)
-        if (config.collapsed) {
-          setCollapsed(false)
-        }
-      } else if (!config.collapsed) {
-        // Start auto-hide when unpinning
-        autoHideTimerRef.current = setTimeout(() => setCollapsed(true), SIDEBAR_AUTO_HIDE_MS)
-      }
-      return next
-    })
-  }
-
-  useEffect(() => () => clearAutoHideTimer(), [clearAutoHideTimer])
-
-  // On mobile, always show expanded view; on desktop, respect collapsed state
-  const isCollapsed = !isMobile && config.collapsed
-
-  // Sidebar width (resizable)
-  const sidebarWidth = isCollapsed ? SIDEBAR_COLLAPSED_WIDTH_PX : (config.width ?? SIDEBAR_DEFAULT_WIDTH_PX)
-
-  // Resize handle state
-  const [isResizing, setIsResizing] = useState(false)
-  const handleResizeStart = (e: React.MouseEvent) => {
-    e.preventDefault()
-    setIsResizing(true)
-    const startX = e.clientX
-    const startWidth = config.width ?? SIDEBAR_DEFAULT_WIDTH_PX
-
-    const handleMouseMove = (moveEvent: MouseEvent) => {
-      const newWidth = Math.min(
-        SIDEBAR_MAX_WIDTH_PX,
-        Math.max(SIDEBAR_MIN_WIDTH_PX, startWidth + (moveEvent.clientX - startX))
-      )
-      setWidth(newWidth)
-    }
-
-    const handleMouseUp = () => {
-      setIsResizing(false)
-      document.removeEventListener('mousemove', handleMouseMove)
-      document.removeEventListener('mouseup', handleMouseUp)
-      document.body.style.cursor = ''
-      document.body.style.userSelect = ''
-    }
-
-    document.body.style.cursor = 'col-resize'
-    document.body.style.userSelect = 'none'
-    document.addEventListener('mousemove', handleMouseMove)
-    document.addEventListener('mouseup', handleMouseUp)
-  }
-
-
-  // Inline rename state for custom sidebar items
-  const [editingItemId, setEditingItemId] = useState<string | null>(null)
-  const [editingName, setEditingName] = useState('')
-
-  // Drag and drop state
-  const [draggedItem, setDraggedItem] = useState<string | null>(null)
-  const [dragOverItem, setDragOverItem] = useState<string | null>(null)
-  const [dragSection, setDragSection] = useState<'primary' | 'secondary' | null>(null)
-  const dragCounter = useRef(0)
-
-  // Cluster status counts — uses the shared helpers from clusters/utils.ts
-  // so that Sidebar, Dashboard, and My Clusters page all agree on counts.
-  // "Unreachable" = reachable===false or confirmed connection error.
-  // "Healthy" = not unreachable and (healthy flag OR has reporting nodes).
-  // "Unhealthy" = everything else, so the three buckets always sum to total.
-  const unreachableClusters = deduplicatedClusters.filter((c) => isClusterUnreachable(c)).length
-  const healthyClusters = deduplicatedClusters.filter((c) => !isClusterUnreachable(c) && isClusterHealthy(c)).length
-  const unhealthyClusters = deduplicatedClusters.length - healthyClusters - unreachableClusters
+  const navSections: NavSection[] = [
+    { id: 'primary', items: config.primaryNav.map(toNavItem) },
+    { id: 'secondary', items: config.secondaryNav.map(toNavItem) },
+  ]
 
   // Handle Add Card click - work with current dashboard
   const handleAddCardClick = () => {
-    // List of dashboards that have card systems
     const cardDashboards = ['/', '/workloads', '/security', '/gitops', '/storage', '/compute', '/network', '/events', '/clusters']
     const currentPath = location.pathname
     const isCustomDashboard = currentPath.startsWith('/custom-dashboard/')
 
     if (cardDashboards.includes(currentPath) || isCustomDashboard) {
-      // Current page has cards - use query param to trigger modal
       if (currentPath === ROUTES.HOME) {
         dashboardContext?.openAddCardModal()
       } else {
-        // Navigate to same page with addCard param to trigger modal
         navigate(`${currentPath}?addCard=true`)
       }
     } else {
-      // On a non-card page, navigate to main dashboard
       dashboardContext?.setPendingOpenAddCardModal(true)
       navigate(ROUTES.HOME)
     }
   }
 
-  // Inline rename handlers - only for user-created custom dashboards (not built-in or template items)
-  const handleDoubleClick = (item: SidebarItem, e: React.MouseEvent) => {
-    if (!item.isCustom || !item.href.startsWith('/custom-dashboard/')) return
-    e.preventDefault()
-    e.stopPropagation()
-    setEditingItemId(item.id)
-    setEditingName(item.name)
-  }
-
-  const handleSaveRename = (itemId: string) => {
-    const trimmed = editingName.trim()
-    if (trimmed) {
-      updateItem(itemId, { name: trimmed })
-      emitDashboardRenamed()
-    }
-    setEditingItemId(null)
-    setEditingName('')
-  }
-
-  // Navigate to clusters page with status filter
-  const handleClusterStatusClick = (status: 'healthy' | 'unhealthy' | 'unreachable') => {
-    navigate(`${ROUTES.CLUSTERS}?status=${status}`)
-  }
-
-  // Handle snoozed card swap restore - navigate to dashboard
-  const handleApplySwap = (_swap: SnoozedSwap) => {
-    // Navigate to main dashboard where the swap can be applied
-    navigate(ROUTES.HOME)
-  }
-
-  // Handle snoozed recommendation restore - navigate to dashboard
-  const handleApplyRecommendation = (_rec: SnoozedRecommendation) => {
-    // Navigate to main dashboard to apply the recommendation
-    navigate(ROUTES.HOME)
-  }
-
-  // Handle snoozed mission restore - navigate to dashboard
-  const handleApplyMission = (_mission: SnoozedMission) => {
-    // Navigate to main dashboard where missions are displayed
-    navigate(ROUTES.HOME)
-  }
-
-  // Drag handlers
-  const handleDragStart = (e: React.DragEvent, itemId: string, section: 'primary' | 'secondary') => {
-    setDraggedItem(itemId)
-    setDragSection(section)
-    e.dataTransfer.effectAllowed = 'move'
-    e.dataTransfer.setData('text/plain', itemId)
-    // Add some delay for visual feedback
-    requestAnimationFrame(() => {
-      const target = e.target as HTMLElement
-      target.style.opacity = '0.5'
-    })
-  }
-
-  const handleDragEnd = (e: React.DragEvent) => {
-    const target = e.target as HTMLElement
-    target.style.opacity = '1'
-    setDraggedItem(null)
-    setDragOverItem(null)
-    setDragSection(null)
-    dragCounter.current = 0
-  }
-
-  const handleDragEnter = (e: React.DragEvent, itemId: string) => {
-    e.preventDefault()
-    dragCounter.current++
-    if (itemId !== draggedItem) {
-      setDragOverItem(itemId)
-    }
-  }
-
-  const handleDragLeave = () => {
-    dragCounter.current--
-    if (dragCounter.current === 0) {
-      setDragOverItem(null)
-    }
-  }
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault()
-    e.dataTransfer.dropEffect = 'move'
-  }
-
-  const handleDrop = (e: React.DragEvent, targetId: string, section: 'primary' | 'secondary') => {
-    e.preventDefault()
-    dragCounter.current = 0
-
-    if (!draggedItem || draggedItem === targetId || section !== dragSection) {
-      setDraggedItem(null)
-      setDragOverItem(null)
-      setDragSection(null)
-      return
-    }
-
-    const items = section === 'primary' ? [...config.primaryNav] : [...config.secondaryNav]
-    const draggedIndex = items.findIndex(item => item.id === draggedItem)
-    const targetIndex = items.findIndex(item => item.id === targetId)
-
-    if (draggedIndex === -1 || targetIndex === -1) return
-
-    // Remove dragged item and insert at target position
-    const [removed] = items.splice(draggedIndex, 1)
-    items.splice(targetIndex, 0, removed)
-
-    // Update order numbers
-    const reorderedItems = items.map((item, index) => ({ ...item, order: index }))
-    reorderItems(reorderedItems, section)
-
-    setDraggedItem(null)
-    setDragOverItem(null)
-    setDragSection(null)
-  }
-
-  const renderIcon = (iconName: string, className?: string) => {
-    const IconComponent = iconRegistry[iconName] as React.ComponentType<{ className?: string }> | undefined
-    return IconComponent ? <IconComponent className={className} /> : null
-  }
-
-  const renderNavItem = (item: SidebarItem, section: 'primary' | 'secondary') => {
-    const isEditing = editingItemId === item.id
-    // Only show the shared Tooltip in collapsed mode — when expanded, the
-    // full item label is already visible, so the tooltip would be redundant.
-    const showTooltip = isCollapsed && !isEditing
-    const tooltipContent = showTooltip
-      ? `${item.name} — ${t('help.sidebarNavItem')}`
-      : ''
-
-    return (
-      <Tooltip
-        key={item.id}
-        content={tooltipContent}
-        side="right"
-        disabled={!showTooltip}
-        wrapperClassName="block w-full"
-      >
-      <div
-        draggable={!isCollapsed && !isEditing}
-        onDragStart={(e) => handleDragStart(e, item.id, section)}
-        onDragEnd={handleDragEnd}
-        onDragEnter={(e) => handleDragEnter(e, item.id)}
-        onDragLeave={handleDragLeave}
-        onDragOver={handleDragOver}
-        onDrop={(e) => handleDrop(e, item.id, section)}
-        className={cn(
-          'group relative transition-all duration-150 w-full',
-          dragOverItem === item.id && dragSection === section && 'before:absolute before:inset-x-0 before:-top-0.5 before:h-0.5 before:bg-purple-500 before:rounded-full',
-          draggedItem === item.id && 'opacity-50'
-        )}
-      >
-        {isEditing ? (
-          // Inline editing mode
-          <div className={cn(
-            'flex items-center gap-3 rounded-lg text-sm font-medium',
-            'bg-purple-500/20 text-purple-400',
-            isCollapsed ? 'justify-center p-3' : 'px-3 py-2'
-          )}>
-            {renderIcon(item.icon, isCollapsed ? 'w-6 h-6' : 'w-5 h-5')}
-            {!isCollapsed && (
-              <input
-                type="text"
-                value={editingName}
-                onChange={(e) => setEditingName(e.target.value)}
-                onBlur={() => handleSaveRename(item.id)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') handleSaveRename(item.id)
-                  if (e.key === 'Escape') { setEditingItemId(null); setEditingName('') }
-                }}
-                autoFocus
-                className="w-[150px] md:w-full md:flex-1 shrink bg-transparent border-b border-purple-500 outline-none text-foreground text-sm min-w-0"
-              />
-            )}
-            {!isCollapsed && <GripVertical className="w-3.5 h-3.5 text-muted-foreground/50 flex-shrink-0" />}
-          </div>
-        ) : (
-          // Normal navigation mode
-          <NavLink
-            to={item.href}
-            onClick={() => emitSidebarNavigated(item.href)}
-            onDoubleClick={(e) => handleDoubleClick(item, e)}
-            onMouseEnter={() => prefetchDashboard(item.href)}
-            className={({ isActive }) => cn(
-              'flex items-center gap-3 rounded-lg text-sm font-medium transition-all duration-200',
-              isActive
-                ? 'bg-purple-500/15 text-purple-400 border-l-[3px] border-purple-500 shadow-[inset_0_0_12px_rgba(168,85,247,0.08)]'
-                : 'text-muted-foreground hover:text-foreground hover:bg-secondary/50 border-l-[3px] border-transparent',
-              isCollapsed ? 'justify-center p-3' : 'px-3 py-2'
-            )}
-            title={item.isCustom && item.href.startsWith('/custom-dashboard/') ? `${item.name} — ${t('sidebar.doubleClickRename')}` : item.name}
-          >
-            {renderIcon(item.icon, isCollapsed ? 'w-6 h-6' : 'w-5 h-5')}
-            {!isCollapsed && (() => {
-              const dashId = HREF_TO_DASHBOARD_ID[item.href]
-              const count = dashId ? DASHBOARD_CONFIGS[dashId]?.cards?.length : null
-              const isGC = isGroundControlItem(item.href)
-              return (
-                <span className="flex-1 min-w-0 flex items-center gap-1">
-                  <span className="truncate">{item.name}</span>
-                  {isGC && (
-                    <Satellite className="w-3.5 h-3.5 text-purple-400 flex-shrink-0" aria-label="Ground Control dashboard" />
-                  )}
-                  {count != null && (
-                    <span className="text-[10px] text-muted-foreground/40 tabular-nums ml-0.5 flex-shrink-0">{count}</span>
-                  )}
-                </span>
-              )
-            })()}
-            {!isCollapsed && (
-              <span className="absolute right-5 top-1/2 -translate-y-1/2 flex items-center gap-1 opacity-100 md:opacity-0 md:group-hover:opacity-100 transition-opacity bg-background/80 backdrop-blur-sm rounded px-1">
-                {!PROTECTED_SIDEBAR_IDS.includes(item.id) && (
-                  <span
-                    role="button"
-                    tabIndex={0}
-                    onClick={(e) => { e.preventDefault(); e.stopPropagation(); removeItem(item.id) }}
-                    onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); e.stopPropagation(); removeItem(item.id) } }}
-                    className="p-1 rounded hover:bg-red-500/20 hover:text-red-400 text-muted-foreground/50 transition-colors"
-                    title={t('sidebar.removeFromSidebar')}
-                    aria-label={t('sidebar.removeFromSidebar')}
-                  >
-                    <X className="w-3.5 h-3.5" aria-hidden="true" />
-                  </span>
-                )}
-                <span
-                  className="p-1 rounded hover:bg-secondary text-muted-foreground/50 hover:text-muted-foreground cursor-grab active:cursor-grabbing transition-colors"
-                  onMouseDown={(e) => e.stopPropagation()}
-                >
-                  <GripVertical
-                    className="w-4 h-4"
-                    aria-hidden="true"
-                  />
-                </span>
-              </span>
-            )}
-          </NavLink>
-        )}
-      </div>
-      </Tooltip>
-    )
-  }
-
   return (
-    <>
-      {/* Mobile backdrop */}
-      {isMobile && config.isMobileOpen && (
-        <div
-          className="fixed inset-0 bg-black/60 backdrop-blur-sm z-overlay md:hidden"
-          onClick={closeMobileSidebar}
-        />
-      )}
-
-      <aside
-        data-testid="sidebar"
-        data-tour="sidebar"
-        onMouseEnter={handleSidebarMouseEnter}
-        onMouseLeave={handleSidebarMouseLeave}
-        className={cn(
-          'fixed left-0 top-16 bottom-0 glass border-r border-border/50 overflow-y-auto scroll-enhanced z-sidebar',
-          !isResizing && 'transition-all duration-300',
-          // Desktop: respect collapsed state
-          !isMobile && (config.collapsed ? 'p-3' : 'p-4'),
-          // Mobile: slide off-screen when closed
-          isMobile && 'p-4',
-          isMobile && !config.isMobileOpen && '-translate-x-full hidden md:flex',
-          isMobile && config.isMobileOpen && 'translate-x-0'
-        )}
-        style={{ width: isMobile ? SIDEBAR_DEFAULT_WIDTH_PX : sidebarWidth }}>
-
-        {/* Primary navigation */}
-        <nav data-testid="sidebar-primary-nav" className="space-y-1">
-          {config.primaryNav.map(item => renderNavItem(item, 'primary'))}
-        </nav>
-
-        {/* Add more dashboards — opens Console Studio at Dashboards section */}
-        {!isCollapsed && (
-          <button
-            onClick={() => {
-              dashboardContext?.openAddCardModal('dashboards')
-            }}
-            className="w-full flex items-center gap-3 px-3 py-1.5 mt-1 text-xs text-muted-foreground/60 hover:text-muted-foreground hover:bg-secondary/30 rounded-lg transition-colors"
-          >
-            <Plus className="w-3.5 h-3.5" />
-            <span>{t('sidebar.addMore', 'Add more...')}</span>
-          </button>
-        )}
-
-        {/* Divider */}
-        <div className="my-6 border-t border-border/50" />
-
-        {/* Secondary navigation */}
-        <nav className="space-y-1">
-          {config.secondaryNav.map(item => renderNavItem(item, 'secondary'))}
-        </nav>
-
-        {/* Snoozed card swaps */}
-        {!isCollapsed && (
-          <div data-tour="snoozed" className="min-w-0">
-            <SnoozedCards
-              onApplySwap={handleApplySwap}
-              onApplyRecommendation={handleApplyRecommendation}
-              onApplyMission={handleApplyMission}
-            />
-          </div>
-        )}
-
-        {/* Add card button */}
-        {!isCollapsed && (
-          <div className="mt-6">
-            <button
-              data-testid="sidebar-add-card"
-              onClick={handleAddCardClick}
-              className="w-full flex items-center justify-center gap-2 px-3 py-2 rounded-lg border border-dashed border-border text-muted-foreground hover:text-foreground hover:border-purple-500/50 hover:bg-purple-500/10 transition-all duration-200"
-            >
-              <Plus className="w-4 h-4" aria-hidden="true" />
-              <span className="text-sm">{t('buttons.addCard')}</span>
-            </button>
-          </div>
-        )}
-
-        {/* Cluster status summary */}
-        {config.showClusterStatus && !isCollapsed && (
-          <div data-testid="sidebar-cluster-status" className="mt-6 p-4 rounded-lg bg-secondary/30">
-            <h4 className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-3">
-              {t('labels.clusterStatus')}
-            </h4>
-            <div className="space-y-2">
-              <button
-                onClick={() => handleClusterStatusClick('healthy')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <CheckCircle2 className="w-3.5 h-3.5 text-green-400" aria-hidden="true" />
-                  {t('labels.healthy')}
-                </span>
-                <span className="text-sm font-medium text-green-400">{healthyClusters}</span>
-              </button>
-              <button
-                onClick={() => handleClusterStatusClick('unhealthy')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <AlertTriangle className="w-3.5 h-3.5 text-red-400" aria-hidden="true" />
-                  {t('labels.unhealthy')}
-                </span>
-                <span className="text-sm font-medium text-red-400">{unhealthyClusters}</span>
-              </button>
-              <button
-                onClick={() => handleClusterStatusClick('unreachable')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <WifiOff className="w-3.5 h-3.5 text-yellow-400" aria-hidden="true" />
-                  {t('labels.offline')}
-                </span>
-                <span className="text-sm font-medium text-yellow-400">{unreachableClusters}</span>
-              </button>
-            </div>
-          </div>
-        )}
-
-        {/* Viewer count + commit hash */}
-        {!isCollapsed && (
-          <div className="mt-auto pt-4 flex flex-col items-center gap-1">
-            <div className="flex items-center justify-center gap-2">
-              <div
-                className="flex items-center gap-1 px-2 text-muted-foreground/60"
-                title={t('sidebar.activeViewers', { count: viewerCount })}
-              >
-                <User className={cn('w-3 h-3', viewersError && 'text-red-400')} />
-                <span className="text-2xs tabular-nums">
-                  {viewersError ? '!' : viewersLoading ? '…' : viewerCount}
-                </span>
-              </div>
-              <span className="text-2xs text-muted-foreground/40 font-mono" title={`Commit: ${__COMMIT_HASH__}`}>
-                {__COMMIT_HASH__.substring(0, 7)}
-              </span>
-            </div>
-            {/* Developer mode: warn when running an older commit, or show upgrade progress */}
-            {channel === 'developer' && hasUpdate && (
-              <div
-                className={cn(
-                  'flex items-center gap-1 text-2xs',
-                  isUpgrading ? 'text-cyan-400/80' : 'text-yellow-400/80',
-                )}
-                title={isUpgrading
-                  ? t('update.upgrading', 'Upgrading...')
-                  : `Behind main — latest: ${latestMainSHA?.substring(0, 7) ?? 'unknown'}`}
-              >
-                {isUpgrading ? (
-                  <Loader2 className="w-3 h-3 animate-spin" />
-                ) : (
-                  <AlertTriangle className="w-3 h-3" />
-                )}
-                <span>
-                  {isUpgrading
-                    ? t('update.upgrading', 'Upgrading...')
-                    : t('sidebar.behindMain', 'Behind main')}
-                </span>
-              </div>
-            )}
-          </div>
-        )}
-      </aside>
-
-      {/* Collapse + Pin controls — top-right of sidebar, above resize handle.
-          Container uses solid `bg-background` and overlaps the sidebar's right
-          border by 1px (see SIDEBAR_CONTROLS_LEFT_OFFSET_PX) so the sidebar's
-          translucent `glass` border doesn't bleed through behind the icons
-          (Issue 8843). */}
-      {!isMobile && !isMissionFullScreen && (
-        <div
-          className="fixed top-[4.5rem] z-sticky flex flex-col gap-1.5 items-center transition-[left] duration-300 bg-background border border-border/50 rounded-full p-1 shadow-md"
-          style={{ left: sidebarWidth + SIDEBAR_CONTROLS_LEFT_OFFSET_PX }}
-        >
-          <button
-            data-testid="sidebar-collapse-toggle"
-            onClick={() => {
-              if (config.collapsed) {
-                // Expanding: also pin so it stays open
-                setCollapsed(false)
-                if (!isPinned) {
-                  setIsPinned(true)
-                  try { localStorage.setItem('sidebar-left-pinned', 'true') } catch { /* ignore */ }
-                  clearAutoHideTimer()
-                }
-              } else {
-                toggleCollapsed()
-              }
-            }}
-            aria-expanded={!config.collapsed}
-            className="hidden md:flex items-center justify-center w-8 h-8 rounded-full bg-secondary text-muted-foreground hover:text-foreground hover:bg-secondary/80 transition-colors"
-            title={config.collapsed ? t('layout.sidebar.expandSidebar') : t('layout.sidebar.collapseSidebar')}
-          >
-            {config.collapsed ? <ChevronRight className="w-3.5 h-3.5" aria-hidden="true" /> : <ChevronLeft className="w-3.5 h-3.5" aria-hidden="true" />}
-          </button>
-          <button
-            onClick={toggleSidebarPin}
-            aria-pressed={isPinned}
-            className={cn(
-              "flex items-center justify-center w-8 h-8 rounded-full transition-colors",
-              isPinned
-                ? "bg-purple-500/15 text-purple-400 hover:bg-purple-500/25"
-                : "bg-secondary text-muted-foreground hover:text-foreground hover:bg-secondary/80"
-            )}
-            title={isPinned ? t('layout.sidebar.unpinSidebar') : t('layout.sidebar.pinSidebar')}
-          >
-            {isPinned ? <Pin className="w-3.5 h-3.5" /> : <PinOff className="w-3.5 h-3.5" />}
-          </button>
-        </div>
-      )}
-
-      {/* Resize handle - drag to adjust sidebar width, starts below pin/collapse buttons */}
-      {!isCollapsed && !isMobile && (
-        <div
-          onMouseDown={handleResizeStart}
-          className={cn(
-            "fixed bottom-0 cursor-col-resize z-sticky hover:bg-purple-500/30 transition-colors hidden md:block",
-            isResizing && "bg-purple-500/50"
-          )}
-          style={{ top: 160, left: sidebarWidth - 3, width: 6 }}
-        />
-      )}
-
-    </>
+    <SidebarShell
+      navSections={navSections}
+      features={{
+        missions: true,
+        addCard: true,
+        addMore: true,
+        clusterStatus: config.showClusterStatus,
+        activeUsers: true,
+        versionCheck: true,
+        dragReorder: true,
+        resize: true,
+        collapsePin: true,
+        snoozedCards: true,
+      }}
+      onAddMore={() => dashboardContext?.openAddCardModal('dashboards')}
+      onAddCard={handleAddCardClick}
+    />
   )
 }
